+++ src/content/amazon/v2/model/AmazonPDPItem.js
@@ -1,0 +1,5 @@
+export class AmazonPDPItem {
+    constructor(asin) {
+        this.asin = asin;
+    }
+}