+++ src/content/amazon/v2/productDetailPage.js
@@ -1,12 +1,19 @@
-import { PRL_COOKIE_NAME_ID, PRL_MISSING_VALUE } from "../../../config/constants";
+import { PRL_MISSING_VALUE } from "../../../config/constants";
+import { AmazonPDPItem } from "./model/AmazonPDPItem";
 import { ProductNavigatorData } from "./model/ProductNavigatorData";
-import { getCookie } from "./util/cookie";
+import { Event } from "./model/Event";
 
-const prolificIDCookieValue = getCookie(PRL_COOKIE_NAME_ID);
-console.log(prolificIDCookieValue);
+// Create the datalayer object, responsible for tracking
+let productNavigatorData = new ProductNavigatorData();
+
+// Emit the "inspect" event
+const asin = document.getElementById('addToCart_feature_div').getAttribute("data-csa-c-asin");
+let item = new AmazonPDPItem(asin);
+let event = new Event(item, "inspect");
+productNavigatorData.pushEvent(event);
 
 try {
-    if (prolificIDCookieValue != PRL_MISSING_VALUE) {
+    if (productNavigatorData.user.id != PRL_MISSING_VALUE) {
         // Disable the add to cart button and show dialog
         document.getElementById("addToCart_feature_div").addEventListener("click", () => {
             alert("Thank you for participating in our survey. You are redirect to the questionaire to answer a few more questions.");
@@ -16,5 +23,3 @@ try {
     }
 } catch (error) { }
 
-// Create the datalayer object, responsible for tracking
-let productNavigatorData = new ProductNavigatorData();